% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{argument}
\alias{argument}
\title{argument class for containing information about argumnets and associated methods}
\description{
The
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{argument$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{argument$new(...)}\if{html}{\out{</div>}}
}

}
}
