% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{createArgument}
\alias{createArgument}
\title{Method for generating a single argument}
\usage{
createArgument(
  sflag,
  lflag,
  name,
  narg,
  flag = FALSE,
  required = TRUE,
  parse.func,
  additional_args = list(),
  parse.func.opt = list()
)
}
\arguments{
\item{sflag}{smallest flag}

\item{lflag}{longest flag}

\item{name}{name of the argument. This becomes the name of any parsed
argument and is used for extracting by name. If null lflag is used instead.}

\item{narg}{argument for deciding the behaviour when passed multiple
arguments. See details for more information.}
}
\description{
This method provides an interface for creating a single
argument. While it is not recommended for use in production ready code,
this allows any user to illustrate the creation of individual parser
components and may potentially speed up interpretation for scripts of 1 - 2
arguments by avoiding the need for an interface to the R6 class. However,
it is still recommended to interact with the standard interface outside of
visualizing the methodology.
}
\details{
For those familiar with the \href{https://docs.python.org/3/library/argparse.html}{argparse}
package in python the \code{narg} argument works in a very similar fashion to
\code{narg} in their package with a few extensions.
The valid options are "?", "*", "+", "N", "N+" and "N-". The first four works
exactly the same, while the latter two provides some addditional options for
"at least" and "at most" N arguments.
\enumerate{
\item \code{narg = "?"} indicates that we accept either 0 or 1 argument.
\item \code{narg = "*"} indicates that any number of arguments (0, 1, 2, ...)
are accepted
\item \code{narg = "+"} indicates that require at least 1 argument.
\item \code{narg = "N"} (where N is an integer) indicates that we accept N and
only N arguments
\item \code{narg = "N+"} indicates that we accept N or more arguments
but not less
\item \code{narg = "N-"} indicates that we accepted N or less arguments.
}

Note that an argument with \code{narg = "+"} and \code{required = FALSE} is
equivalent to specifying \code{narg = "*"}. For \code{narg = "N"} and
\code{narg = "N+"} specifying \code{required = FALSE} will have an effect
similar to \code{"?"} and \code{"+"} but starting from \code{"N"}. Eg. we require either \code{0}
arguments, \code{"N"} or \code{"N+"} arguments.
As a slight note this does mean that \code{narg = 1} is equivalent to \code{narg = "?"}
as the behaviour is fully controlled by \code{required = TRUE/FALSE}.

If the argument is a flag (indicated by setting \code{flag = TRUE}) both \code{narg}
and \code{required} are ignored internally.

FIX ME:: add description of remaining arguments:
}
\note{
if nchar(sflag) > nchar(lflag) the arguments are switched internally.
}
\examples{
# Create a simple argument and extract it's values.
f <- createArgument('h', 'help', 'help', 3, TRUE, FALSE)
getflags(f)
getsflag(f)
getlflag(f)
getnarg(f)
getnargparsed(f)
isflag(f)
isrequired(f)
getname(f)

}
