% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demandReady.R
\name{demandReady}
\alias{demandReady}
\title{Demand that an object is ready for parsing}
\usage{
demandReady(x, ...)
}
\arguments{
\item{x}{a parser object that should be ready for parsing}

\item{...}{further parameters passed to method}
}
\value{
NULL invisible
}
\description{
This function is a simple function to test whether a specific
argument or all arguments in a parser are ready for parsing. This is mainly
thought to be useful for visualization as it is not used internally.
}
