% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_args.R
\name{find_args}
\alias{find_args}
\title{Locate the position of arguments passed from commandline}
\usage{
find_args(args = commandArgs(TRUE), sarg, larg)
}
\arguments{
\item{args}{vector of commandline arguments. Defaults to commandArgs(TRUE)}

\item{sarg}{short argument indicator (one letter arguments). No default}

\item{larg}{long argument indicator (one or more letter).}
}
\value{
An integer vector indicating the positions that started with either
sarg or larg. If both are provided the integer indicates larg over sarg to
avoid any confusion between "-" and "--" (or equivalent).
The vector contains an attribute 'argLen' indicating the end position for
the argument indicator.
}
\description{
This function is intended to find the position of commandline
arguments. Under usual circumstances this is called from other functions but
has been exported for the user. It is intentionally restricted to only take 2
possible argument indicators, but does not in itself restrict the format of
those argument. These restrictions should be checked before calling the
function. The check itself is performed in optimized c++ code.
}
\note{
The function checks whether the argument indicator is followed by an
empty space and throws an error if this is the case. This follows the output
syntax from commandArg, which cannot have "spaces" in arguments unless quoted.

The function does \strong{not} check for other types of invalid arguments
(duplicates, invalid symbol names etc.). These checks should be caught with
pattern matching.
}
\examples{

# Simple argument finding
find_args(c('--abcd', '-d', 'hello', '--ff'), '-', '--')
find_args(c('--abcd', '-d', 'hello', '--ff'), '-')
find_args(c('--abcd', '-d', 'hello', '--ff'), '--')
# Example with help
find_args('-h', '-', '--')
find_args('--help', '-', '--')
# Note that this does not find help:
find_args('--help', '-')
# Example with alternative indicator
find_args('/h', , '/')

\donttest{
# Example of an error:
find_args(c('--abcd ', '-d', 'hello', '--ff'), '-', '--')
}

}
