% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getflags.R
\name{getflags}
\alias{getflags}
\alias{getflags.cmd_arg}
\alias{getflags.cmd_parser}
\title{extract all flags}
\usage{
getflags(x, ...)

\method{getflags}{cmd_arg}(x, ...)

\method{getflags}{cmd_parser}(x, ...)
}
\arguments{
\item{x}{an object of either class cmd_arg or cmd_parser}

\item{...}{further arguments passed the specific method}
}
\value{
If x is a single argument, a character vector of argument names
if x is a cmd_parser, a list is returned with plausible entries "flags",
"positional" and "named" for flag, positional and named arguments
respectively.
}
\description{
This method can be used to extract the flag values of either a
single argument or all arguments provided to a parser. This interface is
primarily thought to be useful for testing and debugging purposes.
}
\section{Methods (by class)}{
\itemize{
\item \code{cmd_arg}: method for class "cmd_arg"

\item \code{cmd_parser}: method for class "cmd_parser"
}}

\seealso{
Other parsers: 
\code{\link{getlflag}()},
\code{\link{getsflag}()}
}
\concept{parsers}
