% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getflags.R
\name{getlflag}
\alias{getlflag}
\title{extract the largest flag}
\usage{
getlflag(x, ...)
}
\arguments{
\item{x}{the object to extract values from (usually cmd_arg or cmd_parser)}

\item{...}{further parameters passed to the specific method}
}
\value{
If x is a single argument, a character vector of argument names
if x is a cmd_parser, a list is returned with plausible entries "flags",
"positional" and "named" for flag, positional and named arguments
respectively.
}
\description{
This method can be used to extract the largst flag within a
single argument or all of the largest flag within a parser. This interface
is primarily thought to be useful for testing and debugging purposes.
}
\seealso{
Other parsers: 
\code{\link{getflags}()},
\code{\link{getsflag}()}
}
\concept{parsers}
