% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getname.R
\name{getname}
\alias{getname}
\title{extract name from argument or parser}
\usage{
getname(x, ...)
}
\arguments{
\item{x}{object to extract the name of}

\item{...}{additional arguments passed to method}
}
\value{
if x is an
}
\description{
extract name from argument or parser
}
