% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnarg.R, R/getparserfun.R,
%   R/getunparsedargs.R, R/parseargs.R
\name{getnarg}
\alias{getnarg}
\alias{getnarg.cmd_arg}
\alias{getnarg.cmd_parser}
\alias{getparsefunc.cmd_arg}
\alias{getparsefunc.cmd_parser}
\alias{getunparsedargs.cmd_arg}
\alias{getunparsedargs.cmd_parser}
\alias{parseargs.cmd_arg}
\alias{parseargs.cmd_parser}
\title{Extract narg from an argument}
\usage{
getnarg(x, ...)

\method{getnarg}{cmd_arg}(x, ...)

\method{getnarg}{cmd_parser}(x, ...)

\method{getparsefunc}{cmd_arg}(x, ...)

\method{getparsefunc}{cmd_parser}(x, ...)

\method{getunparsedargs}{cmd_arg}(x, ...)

\method{getunparsedargs}{cmd_parser}(x, ...)

\method{parseargs}{cmd_arg}(x, ...)

\method{parseargs}{cmd_parser}(x, ...)
}
\arguments{
\item{x}{an object of class \code{cmd_arg} or \code{cmd_parser}}

\item{...}{further parameters passed to method}
}
\value{
if x is an object of \code{cmd_arg} a single character element specifying
the value of narg, otherwise a named vector with \code{narg} for each argument in
the parser
}
\description{
simple method for extracting the values of narg. Mainly thought
to be useful for debugging purposes.
}
\section{Methods (by class)}{
\itemize{
\item \code{cmd_arg}: method for objects of class \code{cmd_arg}

\item \code{cmd_parser}: method for objects of class \code{cmd_parser}

\item \code{cmd_arg}: method for objects of class \code{cmd_arg}

\item \code{cmd_parser}: method for objects of class \code{cmd_parser}

\item \code{cmd_arg}: method for objects of class \code{cmd_arg}

\item \code{cmd_parser}: method for objects of class \code{cmd_parser}

\item \code{cmd_arg}: method for objects of class \code{cmd_arg}

\item \code{cmd_parser}: method for objects of class \code{cmd_parser}
}}

