% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnarg.R
\name{getnargparsed}
\alias{getnargparsed}
\alias{getnargparsed.cmd_arg}
\alias{getnargparsed.cmd_parser}
\title{Extract parsed values for narg}
\usage{
getnargparsed(x, ...)

\method{getnargparsed}{cmd_arg}(x, ...)

\method{getnargparsed}{cmd_parser}(x, ...)
}
\arguments{
\item{x}{an object of class \code{cmd_arg} or \code{cmd_parser}}

\item{...}{further parameters passed to method}
}
\value{
if x is an object of class \code{cmd_arg} a list of length 2 where the
first element indicates the number N and the second indicates comparison
operator \code{<}, \code{<=}, \code{==}, \code{>=} or \code{>}.
}
\description{
simple method for extracing the parsed values of narg. Mainly
useful for debugging purposes.
}
\section{Methods (by class)}{
\itemize{
\item \code{cmd_arg}: method for object class \code{cmd_arg}

\item \code{cmd_parser}: method for object class \code{cmd_parser}
}}

