% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getflags.R
\name{getsflag}
\alias{getsflag}
\alias{getsflag.cmd_arg}
\alias{getsflag.cmd_parser}
\alias{getlflag.cmd_arg}
\alias{getlflag.cmd_parser}
\title{extract the smallest flag}
\usage{
getsflag(x, ...)

\method{getsflag}{cmd_arg}(x, ...)

\method{getsflag}{cmd_parser}(x, ...)

\method{getlflag}{cmd_arg}(x, ...)

\method{getlflag}{cmd_parser}(x, ...)
}
\arguments{
\item{x}{the object to extract values from (usually cmd_arg or cmd_parser)}

\item{...}{further parameters passed to the specific method}
}
\value{
If x is a single argument, a character vector of argument names
if x is a cmd_parser, a list is returned with plausible entries "flags",
"positional" and "named" for flag, positional and named arguments
respectively.
}
\description{
This method can be used to extract the smallest flag within a
single argument or all of the smallest flag within a parser. This interface
is primarily thought to be useful for testing and debugging purposes.
}
\section{Methods (by class)}{
\itemize{
\item \code{cmd_arg}: method for class "cmd_arg"

\item \code{cmd_parser}: method for class "cmd_parser"

\item \code{cmd_arg}: method for class "cmd_arg"

\item \code{cmd_parser}: method for class "cmd_parser"
}}

\seealso{
Other parsers: 
\code{\link{getflags}()},
\code{\link{getlflag}()}
}
\concept{parsers}
