% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getunparsedargs.R
\name{getunparsedargs}
\alias{getunparsedargs}
\title{Extract \code{parser.func} from an argument}
\usage{
getunparsedargs(x, ...)
}
\arguments{
\item{x}{an object of class \code{cmd_arg} or \code{cmd_parser}}

\item{...}{further parameters passed to method}
}
\value{
the return value depends on the value of \code{parser.func.opts}.
The most common types are
\enumerate{
\item Vector, if vectorized = TRUE and input type is vector
\item List if vectorized = TRUE and type is list or vectorized = FALSE
}
}
\description{
simple method for extracting the cmdline arguments passed to a
specific argument prior to being parsed.
}
