% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isflag.R
\name{isflag}
\alias{isflag}
\title{extract flag information from parser or argument}
\usage{
isflag(x)
}
\arguments{
\item{x}{object to extract flag information from.}
}
\value{
if x is an \code{cmd_arg} object, true if the object is a flag argument
otherwise false. If x is an \code{cmd_parser} object then a named logical vector
indicating which arguments are flags and which are not.
}
\description{
this method provides a way indicate whether an argument is a
flag. A flag argument is an argument indicating whether something is true or
false.
}
