% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseargs.R
\name{parseargs}
\alias{parseargs}
\title{Parse and extract parsed values from an argument or parser}
\usage{
parseargs(x, ...)
}
\arguments{
\item{x}{an object of class \code{cmd_arg} or \code{cmd_parser}}

\item{...}{further parameters passed to method}
}
\value{
return values are determined by the
}
\description{
main method for telling the parser to parse its arguments
and return them. Results are saved and is not recomputed in further calls.
}
