% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUnparsedArgs.R
\name{setUnparsedArgs.cmd_arg}
\alias{setUnparsedArgs.cmd_arg}
\title{Method for distributing cmdline arguments to an argument}
\usage{
\method{setUnparsedArgs}{cmd_arg}(x, args, ...)
}
\arguments{
\item{x}{the object to provide unparsed arguments for}

\item{args}{A character vector of arguments to be passed}

\item{...}{further arguments passed to the specific methods}
}
\description{
when distributing arguments to a single argument no parsing is
performed on the argument. Eg. if you have an argument with flags \code{"-f"} and
\code{"--fail"} then the arguments passed are assumed to not contain the flags
themselves. Every element of the string is passed through \code{parser.func}, so
all cleaning needs to be performed before providing the arguments.
}
\seealso{
\link{setUnparsedArgs.cmd_parser}
}
