% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUnparsedArgs.R
\name{setUnparsedArgs.cmd_parser}
\alias{setUnparsedArgs.cmd_parser}
\title{Method for distributing cmdline arguments to a parser containing arguments}
\usage{
\method{setUnparsedArgs}{cmd_parser}(x, args, ...)
}
\arguments{
\item{x}{the object to provide unparsed arguments for}

\item{args}{A character vector of arguments to be passed}

\item{...}{further arguments passed to the specific methods}
}
\description{
when distributing arguments to a parser, parsing is not
instantiated before calling \code{parseUnparsedArgs} on either the parser or the
individual arguments.
}
\seealso{
\link{setUnparsedArgs.cmd_arg}
}
