% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUnparsedArgs.R
\name{setUnparsedArgs}
\alias{setUnparsedArgs}
\title{Methods for distributing cmdline arguments to an argument or parser}
\usage{
setUnparsedArgs(x, args, ...)
}
\arguments{
\item{x}{the object to provide unparsed arguments for}

\item{args}{A character vector of arguments to be passed}

\item{...}{further arguments passed to the specific methods}
}
\description{
These method provides an interface for manual provision of
arguments to either a parser or a specific argument. This action is performed
automatically in normal circumstances. See the individual methods for more
information about how the arguments are handled and specific examples.
These methods are not intended for production code where
\code{parser$parse_args(args)} or \code{parse_args(parser, args)} automates
this action and these methods are provided for visualizing the C++
implementation from R.
}
\seealso{
\link{setunparsedArgs.cmd_arg}, \link{setunparsedArgs.cmd_parser}
}
