% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_args.R
\name{match_args}
\alias{match_args}
\title{Match arguments}
\usage{
match_args(
  cmdArgs = commandArgs(TRUE),
  parserArgs,
  parserArgsRequired,
  helpArg,
  sarg,
  larg
)
}
\arguments{
\item{parserArgs}{arguments to be searched for}

\item{parserArgsRequired}{logical vector indicaing whether an argument is strictly required}

\item{helpArg}{the argument used if one seeks help with the package}

\item{sarg}{short argument indicator (one letter arguments). No default}

\item{larg}{long argument indicator (one or more letter).}
}
\description{
Match arguments
}
\examples{
cmdArgs <- c('--go', 'abc', 'efd', '-g', 'my world', '--garage', 'no', '-f', '-v')
parserArgs <- list(c('-g', '--garage'), '--go', '-f', '-v')
parserArgsRequired <- rep(TRUE, 4)
helpArg <- '--help'
sarg <- '-'
larg <- '--'
cmdline.arguments:::match_args(cmdArgs,
           parserArgs,
           parserArgsRequired,
           helpArg,
           sarg,
           larg)
}
